+++ app/Keycloak/Config.php
@@ -1,0 +1,31 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Keycloak\Dto;
+
+use App\Keycloak\Collection\RealmCollection;
+
+final readonly class Config
+{
+    public string $baseUrl;
+
+    public function __construct(
+        public bool $isEnabled, // This is a feature flag
+        string $baseUrl,
+        public string $clientId,
+        public string $clientSecret,
+        public RealmCollection $realms
+    ) {
+        $this->baseUrl = $this->addTrailingSlash($baseUrl);
+    }
+
+    private function addTrailingSlash(string $uri): string
+    {
+        if (str_ends_with($uri, '/')) {
+            return $uri;
+        }
+
+        return $uri . '/';
+    }
+}
